left and right menu separated
